Changed walks to doubles inside algo and only cast to ints in GUI

+++ src/main/java/com/team18/model/RouteStep.java
@@ -7,7 +7,7 @@ public class RouteStep {
     public boolean walking;
     public double latTo;
     public double lonTo;
-    public int durationMinutes;
+    public double durationMinutes;
     public int startTimeSecondsAfterMidnight;
 
     public String fromStopName;
@@ -18,7 +18,7 @@ public class RouteStep {
     public String headSign;
 
     // Walking constructor
-    public RouteStep(double latTo, double lonTo, int durationMinutes, int startTimeSecondsAfterMidnight, String toStopName) {
+    public RouteStep(double latTo, double lonTo, double durationMinutes, int startTimeSecondsAfterMidnight, String toStopName) {
         this.walking = true;
         this.latTo = latTo;
         this.lonTo = lonTo;
@@ -28,7 +28,7 @@ public class RouteStep {
     }
 
     // Public transit constructor
-    public RouteStep(double latTo, double lonTo, int durationMinutes, int startTimeSecondsAfterMidnight,
+    public RouteStep(double latTo, double lonTo, double durationMinutes, int startTimeSecondsAfterMidnight,
                      String fromStopName, String toStopName, String operatorName, String shortName, String longName, String headSign) {
         this.walking = false;
         this.latTo = latTo;
